tools/notebooklm/requirements.txt
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.0

